+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 - Typescript Template v2
 
 ### Fixes
+- [Issue #3] refval.includes is not a function when referencing non-string variable
 
 ## [1.0.0] Initial Release